In the words of Samuel L Jackson:
I've had it with the MF URLErrors in the MF build!

(added the option to fail on URLErrors in tests. any other errors will still cause test failures)

Also, and not unimportantly, I removed STAR from the default plugins. This was turned on by one of the tests and shipped with 0.7.0 accidentally.