Download annotations from URL provider
When downloading a genome using an url (and setting "url" as provider) with the --annotation flag,
genomepy will assume the link is an FTP directory and seach for a GTF file, or a GFF3 file with a matching genome name.

This system has the same functionality as the other providers.
Resolves #59