Added sanitize annotation feature.
Not all GTFs have the same chromosome/scaffold names as their matching genome.fa.
This feature indexes the mismatched names and replaces them in the GTF.
Additionally, a new BED file is generated from the new GTF.

Also removed sizes from plugins and added it to utils, to match the gaps generation.